requirements.txt
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Deprecated==1.2.18
executing==2.2.0
fastapi==0.115.8
googleapis-common-protos==1.67.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
joblib==1.4.2
logfire==3.5.3
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.2.2
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
redis==5.2.1
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
urllib3==2.3.0
uvicorn==0.34.0
wrapt==1.17.2
zipp==3.21.0
